requirements.txt
# --- Python version ---
# 3.10–3.13 supported with the pins below. (Older stacks used numpy 1.24 + TF 2.15; that does NOT work on 3.13.)
#
# Before install on 3.12+:
#   python -m pip install --upgrade pip
#   python -m pip install "setuptools>=69" "wheel>=0.43"
#   pip install -r requirements.txt
setuptools>=69.0.0
wheel>=0.43.0

# Core ML (NumPy 2.x required for Python 3.13 wheels; TF 2.15 was pinned to NumPy <2 — upgraded below)
numpy>=2.1.0,<3.0.0
pandas>=2.2.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
xgboost>=2.1.0
lightgbm>=4.5.0

# Deep Learning (versions with Py3.13 + NumPy 2 support)
tensorflow>=2.18.0,<2.21.0
# Keras 3 ships with TF 2.16+; separate keras pin not needed for most workflows
torch>=2.5.0

# Time Series
statsmodels>=0.14.2
prophet>=1.1.5
pmdarima>=2.0.4

# Feature Engineering
category_encoders>=2.6.0
feature-engine>=1.6.2

# Model Interpretability
shap>=0.45.0
lime>=0.2.0.1

# Vector Database & Embeddings
pinecone-client>=5.0.0,<8.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.6.1

# LangChain & Agents
langchain>=1.0.0,<2.0.0
langchain-core>=0.3.0,<2.0.0
langchain-community>=0.3.0,<2.0.0
langchain-openai>=0.2.0,<2.0.0
langgraph>=0.2.45,<2.0.0
ollama>=0.1.6

# API & Data Fetching
requests>=2.31.0
aiohttp>=3.9.1
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pymongo>=4.6.1
supabase>=2.4.0,<3.0.0

# MLOps
mlflow>=2.9.2
wandb>=0.16.2
optuna>=3.5.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0.1
joblib>=1.3.2
tqdm>=4.66.1
python-dateutil>=2.8.2
pytz>=2023.3

# Geospatial
geopy>=2.4.1
haversine>=2.8.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
